Feat/incomings expenses

+++ app/Http/Services/CategoryServices.php
@@ -2,9 +2,12 @@
 
 namespace App\Http\Services;
 
-use App\Helpers\ReturnType;
-use App\Helpers\StorageHelper;
+use App\Http\Helpers\FailedData;
+use App\Http\Helpers\ReturnType;
+use App\Http\Helpers\StorageHelper;
+use App\Http\Helpers\SuccessfulData;
 use App\Models\Category;
+use App\Models\Transaction;
 use App\Models\User;
 use Exception;
 use Illuminate\Support\Facades\DB;
@@ -18,38 +21,44 @@ class CategoryServices extends BaseService
         parent::__construct(Category::class);
     }
 
-    public function checkExists($userId, $name): bool
+    public function checkExists(int $userId, string $name): bool
     {
-        if (Category::where(['user_id' => $userId, 'name' => $name])->exists()) {
-            return false;
-        }
-        return true;
+        $isUserCategoryExists = Category::where('user_id', $userId)->where('name', $name)->exists();
+        $isDefaultCategoryExists = Category::whereNull('user_id')->where('name', $name)->exists();
+
+        return ($isUserCategoryExists || $isDefaultCategoryExists);
     }
 
-    public function create(User $user, array $data): ?array
+    public function create(User $user, array $data): object
     {
         try {
             if (!in_array($data['type'], ['incomes', 'expenses'])) {
-                return ReturnType::fail(['type' => "Type is invalid! Available types: [incomes, expenses]"]);
+                return new FailedData('Invalid type', ['type' => "Type is invalid! Available types: [incomes, expenses]"]);
             }
 
-            if (!$this->checkExists($user->id, $data['name'])) {
-                return ReturnType::fail(['name' => "This category name has been used!"]);
+            if ($this->checkExists($user->id, $data['name'])) {
+                return new FailedData('Category name has been used', ['name' => "This category name has been used!"]);
             }
 
             $image = $data['image'];
             $imageUrl = StorageHelper::store($image, "/public/images/categories");
 
-            $categoryData = array_merge($data, ['user_id' => $user->id, 'image' => $imageUrl]);
+            if ($user->hasRole('user')) {
+                $categoryData = array_merge($data, ['user_id' => $user->id, 'image' => $imageUrl, 'default' => false]);
+            }
+            if ($user->hasRole('admin')) {
+                $categoryData = array_merge($data, ['user_id' => null, 'image' => $imageUrl, 'default' => true]);
+            }
+
             $newCategory = $this->createCategory($categoryData);
 
-            return ReturnType::success('Create category successfully!', ['category' => $newCategory]);
+            return new SuccessfulData('Create category successfully!', ['category' => $newCategory]);
         } catch (Exception $error) {
-            return ReturnType::fail($error);
+            return new FailedData('Fail to create category!', ['error' => $error]);
         }
     }
 
-    public function get(User $user, array $inputs): array
+    public function get(User $user, array $inputs): object
     {
         try {
             $type = isset($inputs['type']) ? $inputs['type'] : null;
@@ -57,12 +66,15 @@ class CategoryServices extends BaseService
             $ingoreExists = isset($inputs['ignore_exists']) ?  $inputs['ignore_exists'] : null;
             $month = isset($inputs['month']) ?  $inputs['month'] : null;
             $year = isset($inputs['year']) ?  $inputs['year'] : null;
+            $withPlan = isset($inputs['with_plan']) ?  $inputs['with_plan'] : null;
+            $walletId = isset($inputs['wallet_id']) ?  $inputs['wallet_id'] : null;
 
             $categories = $this->model::query();
 
             if ($ingoreExists && $month && $year) {
                 $categories->where('user_id', $user->id)
-                    ->orWhereNull('user_id')->leftJoin('category_plans', 'categories.id', '=', 'category_plans.category_id')
+                    ->orWhereNull('user_id')
+                    ->leftJoin('category_plans', 'categories.id', '=', 'category_plans.category_id')
                     ->where('category_plans.month', $month)
                     ->where('category_plans.year', $year)
                     ->whereNull('products.category_id')->select('categories.*');
@@ -73,14 +85,19 @@ class CategoryServices extends BaseService
                 ->orWhereNull('user_id')->select('name', DB::raw('count(*) as count'))->groupBy('name')->distinct()->get()->toArray();
 
             $categories = $categories->where('user_id', $user->id)->orWhereNull('user_id')->get()->filter(function ($category) use ($countNames, $default) {
-                if ($default) {
-                    $includeDefault = !is_null($category->user_id);
+                if (!is_null($default)) {
+                    if ($default == true) {
+                        $includeDefault = !is_null($category->user_id);
+                    } else {
+                        $includeDefault = is_null($category->user_id);
+                    }
+                    foreach ($countNames as $name) {
+                        if ($name['name'] == $category->name && $name['count'] > 1 && $includeDefault) return false;
+                    }
                 } else {
-                    $includeDefault = is_null($category->user_id);
-                }
-
-                foreach ($countNames as $name) {
-                    if ($name['name'] == $category->name && $name['count'] > 1 && $includeDefault) return false;
+                    foreach ($countNames as $name) {
+                        if ($name['name'] == $category->name && is_null($category->user_id) && $name['count'] > 1) return false;
+                    }
                 }
 
                 return true;
@@ -102,18 +119,86 @@ class CategoryServices extends BaseService
                 })->values();
             }
 
-            return ReturnType::success("Get categories successfully", ['categories' => $categories]);
+            if ($withPlan && $month && $year) {
+                $categories = $categories->map(function ($category) use ($user, $month, $year, $walletId) {
+                    $plan = app(CategoryPlanService::class)->get($user, ['month' => $month, 'year' => $year, 'category_id' => $category->id, 'wallet_id' => $walletId]);
+
+                    if ($plan instanceof SuccessfulData) {
+                        $category = (object) array_merge($category->toArray(), ['plan' => $plan->getData()['plans']->first()]);
+                    } else {
+                        $category = (object) array_merge($category->toArray(), ['plan' => null]);
+                    }
+
+                    return $category;
+                });
+            }
+
+            return new SuccessfulData("Get categories successfully", ['categories' => $categories]);
         } catch (Exception $error) {
-            return ReturnType::fail($error);
+            return new FailedData('Failed to get categories', ['error' => $error]);
         }
     }
 
-    public function update($data, int $id): array
+    public function getDefault(array $inputs): object
+    {
+        try {
+            $type = isset($inputs['type']) ? $inputs['type'] : null;
+            $search = isset($inputs['search']) ? $inputs['search'] : null;
+
+            $categories = Category::where('default', 1);
+            if ($type) {
+                $categories->where('type', $type);
+            }
+
+            if ($search) {
+                $categories->where('name', 'LIKE', "%" . $search . "%");
+            }
+
+            $categories = $categories->get()->map(function ($category) {
+
+                $transactionsCount = Transaction::whereHas('category', function ($query) use ($category) {
+                    $query->where('name', $category->name);
+                })->count();
+
+                $usersCount = User::whereHas('transactions', function ($query) use ($category) {
+                    $query->whereHas('category', function ($query) use ($category) {
+                        $query->where('name', $category->name);
+                    });
+                })
+                    ->count();
+
+                $category->transactions_count = $transactionsCount;
+                $category->users_count = $usersCount;
+                return $category;
+            });
+
+            return new SuccessfulData('Get default categories successfully!', ['categories' => $categories]);
+        } catch (Exception $error) {
+            return new FailedData('Can not get default categories!', ['error' => $error]);
+        }
+    }
+
+    public function getDefaultCount(): object
+    {
+        try {
+            $count = Category::where('default', 1)->get()->count();
+
+            return new SuccessfulData('', ['count' => $count]);
+        } catch (Exception $error) {
+            return new FailedData('Failed to get count of default categories', ['error' => $error]);
+        }
+    }
+
+    public function update($data, int $id): object
     {
         try {
             $category = $this->getById($id);
             if (!$category) {
-                return ReturnType::fail('Category not found!');
+                return new FailedData('Category not found!', ['error' => 'category']);
+            }
+
+            if ($category->default == 0 && $this->checkExists($category->user_id, $data['name'])) {
+                return new FailedData("This category name has been used!", ['name' => "This category name has been used!"]);
             }
 
             $image = isset($data['image']) ? $data['image'] : null;
@@ -132,32 +217,59 @@ class CategoryServices extends BaseService
             $updated = $category->update($data);
 
             if (!!!$updated) {
-                return ReturnType::fail('Update fails or category not found!');
+                return new FailedData('Update fails or category not found!');
             }
 
-            return ReturnType::success('Update category successfully!');
+            return new SuccessfulData('Update category successfully!');
         } catch (Exception $error) {
-            return ReturnType::fail($error);
+            return new FailedData($error);
         }
     }
 
-    public function delete($id): array
+    public function delete(int $id): object
     {
         try {
             $category = $this->getById($id);
 
+            if (!$category) {
+                return new FailedData('Category not found!');
+            }
+
+            // Delete all plans belong to this cate
+            app(CategoryPlanService::class)->deleteByCategoryId($category->id);
+
             // Delete all transactions belongs to the category
-            foreach ($category->transactions as $transaction) {
-                if ($transaction->image) {
-                    $imagePath = Str::after($transaction->image, '/storage');
+            if ($category->default == 0) {
+                foreach ($category->transactions as $transaction) {
+                    if ($transaction->image) {
+                        $imagePath = Str::after($transaction->image, '/storage');
+                        StorageHelper::delete($imagePath);
+                    }
+
+                    app(TransactionServices::class)->delete($transaction->id);
+                }
+
+                if ($category->image) {
+                    $imagePath = Str::after($category->image, '/storage');
                     StorageHelper::delete($imagePath);
                 }
 
-                app(TransactionServices::class)->delete($transaction->id);
+                $deleted = $this->model::destroy($id);
             }
 
+            return new SuccessfulData('Delete category successfully!');
+        } catch (Exception $error) {
+            return new FailedData('Failed to delete category!');
+        }
+    }
+
+    public function deleteDefault(int $id): object
+    {
+        try {
+            $category = $this->getById($id);
+
             if (!$category) {
-                return ReturnType::fail('Category not found!');
+                return new FailedData('Category not found!');
             }
 
             if ($category->image) {
@@ -167,12 +279,12 @@ class CategoryServices extends BaseService
 
             $deleted = $this->model::destroy($id);
             if (!$deleted) {
-                return ReturnType::fail('Delete fails or category not found!');
+                return new FailedData('Delete fails or category not found!');
             }
 
-            return ReturnType::success('Delete category successfully!');
+            return new SuccessfulData('Delete category successfully!');
         } catch (Exception $error) {
-            return ReturnType::fail($error);
+            return new FailedData('Failed to delete default category!');
         }
     }
 
@@ -186,13 +298,24 @@ class CategoryServices extends BaseService
         return Category::find($id);
     }
 
-    public function getWithSameName(int $id, string $name): ?Category
+    public function getWithSameNameOfUser(int $userId, int $id, string $name): ?Category
     {
-        return Category::where('name', $name)->where('id', '!=', $id)->first();
+        return Category::where(['user_id' => $userId, 'name' => $name])->where('id', '!=', $id)->first();
     }
 
-    public function createCategory($data)
+    public function createCategory(array $data): ?Category
     {
         return $this->model::create($data);
+    }
+
+    public function createBasedOnDefault(int $userId, Category $defaultCategory): ?Category
+    {
+        return Category::create([
+            'name' => $defaultCategory->name,
+            'type' => $defaultCategory->type,
+            'image' => $defaultCategory->image,
+            'user_id' => $userId,
+            'default' => false
+        ]);
     }
 }
